requirement.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-paypal==2.0
django-widget-tweaks==1.4.12
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
